eq
This is a script I created using bash to act as a calculator on the terminal. It takes two possible initial arguments one being '-m' and the other being '-gamma' which I made purely for kicks because it calculates the gamma by giving it the velocity which saves a lot of time. However just using the -m parameter and then inputting the calculation you want. So an example of the terminal command would be the following...  

'eq -m 1+1'

The script can take multiple arguments using parentheses. 

'eq -m "1+1" "1+5"'

Warnings:
Does not take any spaces.
There is a problem with displaying a zero because I can't seem to figure out how to get rid of it. I'll hopefully get rid of it later.