A graphical version of eq
This uses zenity to make a window asking for a password. Which you can change to anything but I made it 'terminal'. Simply changing the line 9 will allow you to change the password. this requires the eq file for it to work. You have to sudo chmod +x interface this in your /usr/bin file for it to give access as well like all the other bash files.